Added ivory-operation sub-project.
This is to consolidate ivory-ingest, ivory-extract and ivory-tools,
moving towards some of the larger code organisation clean-ups.

This the first step - I will move each piece individually, so they
can be merged / managed easier as we go.

The name - ivory-operation isn't ideal in that it doesn't roll
off the tounge, but it is at least consistent with noun theme
of -storage, -data, -core etc...