First move chunk to consolidate things under ivory-operation.
The 'operation' name is pretty unpleasant, so now would be
a good time for someone to come up with something better
if possible.

This follows #175.

Ideally this would of just been ivory-tools as a first step,
but as thing turned out things were _very_ tangled up via
interdependencies on tests (which validates this consolidation
in a sense). So this change includes ivory-tools and
ivory-extract changes.