Remove ivory-generate.
It is not complete, it no longer works properly.

It should be revisited in the future (but probably
in a separate repository) with a better design
that would facilitate generation of more realistic
datasets (quality and scale wise). And should
also leverage what ivory does with it's testing
arbitraries which were almost 100% duplicated
within this code.