Chance pivot reduction size calculation.
The larger justification for this is that we are going from the
far more efficient thrift representation into strings. So there
is an explosion of size. With this set to 1.gb it resulted in
too much being sent to each reducer. The magic number isn't
great, but at least this one should work a bit better.